requirements.txt
matplotlib==3.8.2
numpy==1.24.4
pandas==2.1.4
PyYAML==6.0
